# app/static/swagger.yaml
swagger: '2.0'
info:
  version: "0.0.1"
  title: "Mechanic shop API"
  description: API documentation for Mechanic shop
host: "127.0.0.1:5000"
schemes:
  - "http"
consumes:
  - "application/json"
produces:
  - "application/json"
securityDefinitions:
  bearerAuth:
    type: "apiKey"
    name: "Authorization"
    in: "header"
paths:
...
  /inventory:
    post:
      tags:
        - "inventory_items"
      summary: "Create a new inventory item"
      description: "Create a new inventory item in the system."
      parameters:
        - in: "body"
          name: "Body"
          description: "Inventory item object that needs to be added"
          required: true
          schema:
            type: object
            properties:
              name:
                type: string
                example: "Oil Filter"
              quantity:
                type: integer
                example: 50
              price:
                type: number
                format: float
                example: 15.99
      responses:
        201:
          description: "Inventory item created successfully"
          schema:
            type: object
            properties:
              id:
                type: integer
              name:
                type: string
              quantity:
                type: integer
              price:
                type: number
                format: float
          examples:
            application/json:
              {
                "id": 1,
                "name": "Oil Filter",
                "quantity": 50,
                "price": 15.99
              }
        400:
          description: "Invalid input"
        500:
          description: "Server error"
    get:
      tags:
        - "inventory_items"
      summary: "Get all inventory items"
      description: "Retrieve a list of all inventory items in the system."
      responses:
        200:
          description: "A list of inventory items"
          schema:
            type: array
            items:
              type: object
              properties:
                id:
                  type: integer
                name:
                  type: string
                quantity:
                  type: integer
                price:
                  type: number
                  format: float
          examples:
            application/json:
              [
                {
                  "id": 1,
                  "name": "Oil Filter",
                  "quantity": 50,
                  "price": 15.99
                },
                {
                  "id": 2,
                  "name": "Brake Pads",
                  "quantity": 30,
                  "price": 45.99
                }
              ]
        500:
          description: "Server error" 
    put:
      tags:
        - "inventory_items"
      summary: "Update an inventory item"
      description: "Update an existing inventory item's information."
      parameters:
        - in: "body"
          name: "Body"
          description: "Inventory item object that needs to be updated"
          required: true
          schema:
            type: object
            properties:
              name:
                type: string
                example: "Oil Filter"
              quantity:
                type: integer
                example: 50
              price:
                type: number
                format: float
                example: 15.99
          responses:
          200:
            description: "Inventory item updated successfully"
            schema:
              type: object
              properties:
                id:
                  type: integer
                name:
                  type: string
                quantity:
                  type: integer
                price:
                  type: number
                  format: float
            examples:
              application/json:
                {
                  "id": 1,
                  "name": "Oil Filter",
                  "quantity": 50,
                  "price": 15.99
                }
    delete:
        tags:
          - "inventory_items"
        summary: "Delete an inventory item"
        description: "Delete an existing inventory item from the system."
        parameters:
          - in: "path"
            name: "id"
            type: integer
            required: true
            description: "ID of the inventory item to delete"
        responses:
        200:
          schema:
            type: object
            properties:
              message:
                type: string
          description: "Inventory item deleted successfully"
          examples:
            application/json:
              {
                "message": "Inventory item deleted successfully!"
              }
    get/search:
        tags:
          - "inventory_items"
        summary: "Search inventory items by name"
        description: "Search for inventory items by their name."
        parameters:
          - in: "query"
            name: "name"
            type: string
            required: true
            description: "Name of the inventory item to search for"
        responses:
          200:
            description: "A list of matching inventory items"
            schema:
              type: array
              items:
                type: object
                properties:
                  id:
                    type: integer
                  name:
                    type: string
                  quantity:
                    type: integer
                  price:
                    type: number
                    format: float
            examples:
              application/json:
                [
                  {
                    "id": 1,
                    "name": "Oil Filter",
                    "quantity": 50,
                    "price": 15.99
                  }
                ]
          400:
            description: "Invalid input"
          500:
            description: "Server error" 
...